refac soovqe qiskit
