+++ qc2/algorithms/algorithms_results.py
@@ -47,34 +47,46 @@ class VQEResults(VQEBASEResults):
         super().__init__()
 
 
-class OOVQEResults(VQEResults):
+class SAOOVQEResults():
     """Extends VQEResults to provide a results data container for oo-VQE."""
-    def __init__(self) -> None:
+    def __init__(self, 
+                 state_weights=None, 
+                 energy=None,
+                 circuit_parameters=None, 
+                 orbital_parameters=None) -> None:
         super().__init__()
-        self._optimal_circuit_params: Optional[List] = None
-        self._optimal_orbital_params: Optional[List] = None
-        self._circuit_parameters: Optional[List] = None
-        self._orbital_parameters: Optional[List] = None
+        self._energy: Optional[List] = energy
+        self._circuit_parameters: Optional[List] = circuit_parameters
+        self._orbital_parameters: Optional[List] = orbital_parameters
+        self._state_weights: Optional[List] = state_weights
+        self._optimal_phase: Optional[float] = None
+
+
+    @property
+    def optimal_energy(self) -> Optional[float]:
+        """Returns optimal energy."""
+        return self._energy[-1][0]
+
+    @property
+    def energy(self) -> Optional[List]:
+        """Returns list of energies of all iterations."""
+        return self._energy
+
+    @energy.setter
+    def energy(self, value: List) -> None:
+        """Sets list of energies of all iterations."""
+        self._energy = value
 
     @property
     def optimal_circuit_params(self) -> Optional[List]:
         """Returns the optimal circuit parameters."""
-        return self._optimal_circuit_params
-
-    @optimal_circuit_params.setter
-    def optimal_circuit_params(self, value: List) -> None:
-        """Sets optimal circuit parameters."""
-        self._optimal_circuit_params = value
+        return self._circuit_parameters[-1]
 
     @property
     def optimal_orbital_params(self) -> Optional[List]:
         """Returns optimal orbital parameters."""
-        return self._optimal_orbital_params
+        return self._orbital_parameters[-1]
 
-    @optimal_orbital_params.setter
-    def optimal_orbital_params(self, value: List) -> None:
-        """Sets optimal orbital parameters."""
-        self._optimal_orbital_params = value
 
     @property
     def circuit_parameters(self) -> Optional[List]:
@@ -96,18 +108,36 @@ class OOVQEResults(VQEResults):
         """Sets list with parameters of all iteration."""
         self._orbital_parameters = value
 
-class SAOOVQEResults(OOVQEResults):
-    """Extends OOVQEResults to provide a results data container for state-averaged oo-VQE."""
-    def __init__(self) -> None:
-        super().__init__()
-        self._optimal_state_weights: Optional[List] = None
+    @property
+    def state_weights(self) -> Optional[List]:
+        """Returns optimal state weights."""
+        return self._state_weights
+
+    @state_weights.setter
+    def state_weights(self, value: List) -> None:
+        """Sets optimal state weights."""
+        self._state_weights = value
 
     @property
-    def optimal_state_weights(self) -> Optional[List]:
-        """Returns optimal state weights."""
-        return self._optimal_state_weights
+    def optimal_phase(self) -> Optional[float]:
+        """Returns optimal phase."""
+        return self._optimal_phase
 
-    @optimal_state_weights.setter
-    def optimal_state_weights(self, value: List) -> None:
-        """Sets optimal state weights."""
-        self._optimal_state_weights = value 
+    @optimal_phase.setter
+    def optimal_phase(self, value: float) -> None:
+        """Sets optimal phase."""
+        self._optimal_phase = value
+
+    def update(self, theta: List, kappa: List, energy: List) -> None:
+        """
+        Updates the internal state with new circuit parameters, orbital parameters, 
+        and energy values.
+
+        Args:
+            theta (List): New set of circuit parameters to be appended.
+            kappa (List): New set of orbital parameters to be appended.
+            energy (List): New energy values to be appended.
+        """
+        self._circuit_parameters.append(theta)
+        self._orbital_parameters.append(kappa)
+        self._energy.append(energy)
