+++ qc2/algorithms/qiskit/__init__.py
@@ -1,9 +1,9 @@
 """qc2 algorithms package for qiskit."""
-from .vqe import VQE
-from .oo_vqe import OO_VQE
-from .sa_oo_vqe import SA_OO_VQE
-from .qpe import QPE
-from .iqpe import IQPE
+from .vqe.vqe import VQE
+from .vqe.oo_vqe import OO_VQE
+from .vqe.sa_oo_vqe import SA_OO_VQE
+from .qpe.qpe import QPE
+from .qpe.iqpe import IQPE
 
 __all__ = [
     "VQE",
